fix: socperf interface LimitRequest adapt int64_t int type

+++ services/native/src/thermal_action/action/action_soc/action_cpu_lit.cpp
@@ -94,7 +94,7 @@ void ActionCpuLit::Execute()
 int32_t ActionCpuLit::CpuRuquest(uint32_t freq)
 {
     std::vector<int32_t> tags;
-    std::vector<int32_t> configs;
+    std::vector<int64_t> configs;
 
     if (!g_service->GetSimulationXml()) {
         tags.push_back(LIM_CPU_LIT_ID);
